RST Efficiency Summary Data

Dataset containing summarized efficiency monitoring data for all SR JPE tribuatary and mainstem monitoring programs

efficiency_summary

Format

date_released: Date that efficiency trial release occurs (format - %Y-%m-%d)

release_id: The unique identifier for each release trial

stream: Which stream the RST is located on

site: Site name

subsite: Name of trap

site_group: Site group, used to seperate traps within the same stream that have unique environmental conditions.

number_released: Count of fish released

number_recaptured: Count of fish recaptured in RST on a specific recapture date, when NA trap is not fished

run: Run of fish released

life_stage: Lifestage of fish released

origin: Fish origin (natural, hatchery, mixed, unknown, not recorded, or NA)

Source

Raw datasets and original documentation of data processing and QC process can be found for each stream using the links below. Data is combined here

Battle Creek: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.

Butte Creek: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.

Clear Creek: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.

Deer Creek: Data on EDI

Feather River: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.

Mill Creek: Data on EDI

Knights Landing: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.

Tisdale: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.

Delta Entry: Data on EDI.

Yuba River: Data on EDI, Original Data Exploration Script. Additional data exploration scripts are located in GitHub.