Feather Hatchery Release

Format

release_location_name: Name of the location where fish were released ("FEATHER AT GRIDLEY", "FEATHER AT LIVE OAK", "FEATHER AT YUBA CITY", "FEATHER BEL THRM HI FLOW", "FEATHER BOYDS PUMP RAMP", "FEATHER R HATCHERY")

avg_weight: Average weight of fish released

avg_length: Average length of fish released

first_release_data: First date of release. Typically first and last release date are the same though some releases happen over days/weeks

last_release_data: Last date of release

date_span: Number of days between the first and last release date

mid_release_data: Midpoint date between first and last release date

delta_distance: Distance to Delta Entry (1st St Bridge)

release_latitude: Latitude of the release location

release_longitude: Longitude of releate location

group_tagcode: Tag codes or IDs grouped together

group_total_marked_N: Total number of marked fish included in the release

group_total_unmarked_N: Total number of unmarked fish included in the release

group_total_release_N: Total number of fish released

group_mark_rate: Mark rate of fish released

month: Month of first release date used to join environmental covariates

year: Water year (Oct-Sep) of first release date used to join environmental covariates

exceedance_flow_year_type: Exceedance year type. This covariate is available in forecast_covariates (https://srjpe.github.io/SRJPEdata/articles/forecast_covariates.html)

monthly_max_flow: Monthly max flow. This covariate is available in forecast_covariates (https://srjpe.github.io/SRJPEdata/articles/forecast_covariates.html)

feather_hatchery_release This table summarizes coded wire tag (CWT) data from Feather River Hatchery releases by release group. Data source: Regional Mark Processing Center: https://www.rmpc.org/ filtered to Feather River Hatchery releases datasets