Make Fish Encounter History

Create an encounter history dataframe based on detections and aggregated receiver metadata.

make_fish_encounter_history(
  detections,
  aggregated_reciever_metadata,
  released_fish_table
)

Arguments

detections

A dataframe containing detections information.

aggregated_receiver_metadata

A dataframe containing aggregated receiver metadata.

Value

An encounter history dataframe. A matrix of every fish tagged for a given studyID at every given receiver site (that is in reach.meta.aggregate) and whether it was present 1 or absent 0 in the detection dataframe.

Details

This function creates an encounter history dataframe, which is a matrix of every fish tagged for a given studyID at every given receiver site (that is in reach.meta.aggregate), indicating whether the fish was present (1) or absent (0) in the detection dataframe. It utilizes the earliest detection time for each fish at each receiver site to determine presence or absence.

Examples

# Create encounter history dataframe
encounter_history <- make_fish_encounter_history(detections_data, aggregated_receiver_metadata)
#> Error in mutate(detections %>% filter(receiver_general_location %in% aggregated_reciever_metadata$receiver_general_location) %>%     group_by(fish_id, receiver_general_location, receiver_general_river_km) %>%     summarise(min_time = min(time, na.rm = TRUE)) %>% arrange(fish_id,     min_time), detect = 1): could not find function "mutate"