Passage to Spawner Model Covariates Standard

The passage to spawner model covariates standard table contains environmental covariates for use in the Passage to Spawner (P2S) model. The environmental variables are standardized within a column to center around 0. Selection and preparation of the covariates are detailed in vignette("prep_environmental_covariates.Rmd", package = "SRJPEdata").

p2s_model_covariates_standard

Format

A tibble with 192 rows and 7 columns

year: year

stream: stream associated with covariate data

wy_type: water year type, binary variable describing wet(1) or dry(0)

max_flow_std: standardized maximum flow data for adult migration period (March - August)

gdd_std: standardized cumulative degree days above 20 degrees Celsius for adult migration (Sacramento March - May) and holding (Tributaries May - August)

passage_index: standardized total adult returns (not currently utilized by model)

median_passage_timing_std: standardized median return week (not currently utilized by model)