Redd Survey Data

Raw redd survey monitoring data. Redd survey data is available for Battle Creek, Clear Creek, Feather River, Mill Creek, Yuba River.

redd

Format

date: year data were collected

latitude: latitude of redd location

longitude: longitude of redd location

reach: survey reach where data were collected

river_mile: river mile of redd location

redd_id: unique identifier of redd

fish_guarding: whether or not fish were observed guarding the redd

redd_measured: whether or not the redd was measured

redd_width: widdth of redd (m)

redd_length: length of redd (m)

velocity: measured stream velocity (ft/s)

age: age of redd

age_index: number of times a unique redd was aged If 0, the redd was surveyed but not aged.

redd_count: number of redds surveyed

stream: stream data were collected on

year: year data were collected

survey_method: surveying method

run: run of fish associated with redd

species: species of fish associated with redd

depth_m: depth of water (m)

starting_elevation_ft: starting elevation (ft)

num_of_fish_on_redd: number of fish observed on redd

redd_substrate_class: size class of substrate determined by millimeter size

tail_substrate_class: size class of tail substrate determined by millimeter size

pre_redd_substrate_class: size class of substrate pre-redd determined by millimeter size

Source

Redd data were provided by monitoring programs. These data are currently being uploaded to EDI. In the meantime you can find data exploration scripts for each tributary here and combined redd data here.