Weekly Hours Fished

Datasets containing the hours fished for each stream, site, and week. See vignette("trap_effort", package = "SRJPEdata") for more details.

weekly_hours_fished

Format

Dataframe containing 7 columns (stream, site, subsite, site_group, week, year, hours_fished)

year: year

stream: Stream RST is located on

site: Site RST is located on

site_group: Site group

week: Week

hours_fished: Number of hours that a trap was operated in a given week. Capped at 168.